### 2023-01-12 #5071 @editorialbot
[PRE REVIEW]: SATLLA0: A Flight Software Platform for Aerospace and STEM Education
**Submitting author:** @ronyr (<a href="http://orcid.org/0000-0002-1300-5236">Rony Ronen</a>)
**Repository:** https://github.com/kcglab/satllazero
**Branch with paper.md** (empty if default branch):
**Version:** 4.0.5
