### 2023-01-12 c286086
Fix missing DOI and wrong orcid id.
