+++ paper/paper.md
@@ -12,7 +12,7 @@ authors:
   - name: Michael Britvin
     affiliation: "2"
   - name: Boaz Ben Moshe
-    orcid: 00000-0002-1580-5421
+    orcid: 0000-0002-1580-5421
     affiliation: "1"
 affiliations:
   - name: School of Computer Science, Ariel University, 47100, Israel
