+++ paper/paper.bib
@@ -14,6 +14,7 @@
   pages={126409--126429},
   year={2019},
   publisher={IEEE}
+  doi={10.1109/ACCESS.2019.2927931},
 }
 
 @article{miranda2019comparative,
@@ -23,4 +24,5 @@
   volume={11},
   year={2019},
   publisher={SciELO Brasil}
+  doi={10.5028/jatm.v11.1081},
 }
