+++ paper/paper.bib
@@ -13,8 +13,8 @@
   volume={7},
   pages={126409--126429},
   year={2019},
-  publisher={IEEE}
-  doi={10.1109/ACCESS.2019.2927931},
+  publisher={IEEE},
+  doi={10.1109/ACCESS.2019.2927931}
 }
 
 @article{miranda2019comparative,
@@ -23,6 +23,6 @@
   journal={Journal of Aerospace Technology and Management},
   volume={11},
   year={2019},
-  publisher={SciELO Brasil}
-  doi={10.5028/jatm.v11.1081},
+  publisher={SciELO Brasil},
+  doi={10.5028/jatm.v11.1081}
 }
