### 2023-02-11 #5147 @editorialbot
[REVIEW]: SATLLA0: A Flight Software Platform for Aerospace and STEM Education
**Submitting author:** @ronyr (<a href="http://orcid.org/0000-0002-1300-5236">Rony Ronen</a>)
**Repository:** https://github.com/kcglab/satllazero
**Branch with paper.md** (empty if default branch):
**Version:** 4.0.13
