### 2023-02-12 #5147 @ronyrsim
@federeghe thanks for your review. This project is mostly done by me and Michael and is supervised by Prof. Boaz Ben-Moshe. I'm leading the project and responsible for the s/w while Michael is responsible for the (h/w)(https://github.com/kcglab/satllazero/tree/main/docs/hardware) and the assembly. Thanks.
