### 2023-02-12 kcglab/satllazero#14 @federeghe
Repository documentation and structure
This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/5147

I think that the README should be improved, especially if the repository is meant for students. Opening the repository without the paper provides no information of what the repository contains (hardware? software? an operating systems? a set of on-board tasks? etc.).

For instance, the README is missing:
- [x] Detailed statements on what the repository contains
- [x] A clear list of features
- [x] The list of supported boards/processors (From the README seems Arduino, but it is actually a Raspberry the OBC)
- [x] What is Master/OBC/Ground

I would also revise the directory structure of the repository. I don't expect to find a `docs` directory but the documentation is elsewhere (Wiki + *.md files in the root). The `docs` contains, instead, the structural design and hardware files. My suggestion is to re-organize the directory structure, something like: `hardware/`, `software/`, `stl/`, `paper/`, etc. These are only the first words popped up in my mind, feel free to choose any names you like but I suggest a more structured directory tree.
