### 2023-02-12 da6024d
#14 Repository documentation and structure
