### 2023-02-13 bbd0184
#14 Repository documentation and structure. Add Features.
