### 2023-02-27 #5147 @federeghe
@ronyr Regarding the paper, "statement of need" says that existing open source software didn't attract much attention and they have not been reused. Instead, you provide this software with that _mission_. My question is: what are the features of your software that differentiate it from the other existing software (that failed)? I think that the answer to this question should be the key point of the statement of need.
