+++ paper/paper.md
@@ -33,8 +33,10 @@ SATLLA-0 is divided into three main libraries, with the main library, SAT0_Maste
 
 
 # Statement of need
-Building a nanosatellite and developing flight software is challenging for newcomers to the space (i.e., new-space) field due to the lack of literature and high cost. There are open-source frameworks for space projects developed by space agencies, universities, or commercial companies to serve as reference designs and encourage code reuse in their future projects [@jamie2017basic]. These frameworks have not yet gained significant adoption outside their host organizations.
-
+Building a nanosatellite and developing flight software is challenging for newcomers to the space (i.e., new-space) field due to the lack of literature and high cost. There are open-source frameworks for space projects developed by space agencies, universities, or commercial companies to serve as reference designs and encourage code reuse in their future projects [@jamie2017basic]. These frameworks have not yet gained significant adoption outside of their host organizations [@miranda2019comparative] for the following reasons:
+- Flight software is tailored to a satellite of a specific design and is not transferable to another design.
+- Flight software is not based on open source software or the entire flight software has not been released as open source software.
+- Flight software is too complicated and may not be suitable for organizations just starting out or for STEM Education.
 
 At the same time, using an existing software framework leads to a shorter development program at lower costs and better quality. These effects are beneficial for organizations such as universities or schools that want to enter the space field. These reasons have led us to release the SATLLA-2B flight software as open-source code with additional documents to build a nanosatellite.
 
