### 2023-02-28 #5147 @ronyronen
@federeghe Thanks for your comment. I've updated the PAPER and published a new release: https://github.com/kcglab/satllazero/releases/tag/4.0.9
