### 2023-04-17 #5147 @federeghe
@prashjha I'm trying to find an Arduino to test the code. The software doesn't have a continuos integration, so I would like to try to check that at least compile and start. If the authors (@ronyronen) want to add the CI, it can surely speedup the process.
