+++ .github/workflows/arduino_ci.yml
@@ -1,63 +1,57 @@
-# This is the name of the workflow, visible on GitHub UI.
-name: Arduino Build
+name: build
 
-# Controls when the action will run.
-on:
-  # Triggers the workflow on push or pull request events but only for the main branch
-  push:
-    branches: [ dev ]
-  pull_request:
-    branches: [ dev ]
+on: [push, pull_request]
 
-# This is the list of jobs that will be run concurrently.
-# Since we use a build matrix, the actual number of jobs
-# started depends on how many configurations the matrix
-# will produce.
+env:
+  IDE_VERSION: 1.8.19
+  TEENSY_VERSION: 158
+  IDE_LOCATION: /usr/local/share/arduino
+
 jobs:
-  # This is the name of the job - can be whatever.
   build:
-    # Here we tell GitHub that the jobs must be determined
-    # dynamically depending on a matrix configuration.
-    strategy:
-      matrix:
-        # The matrix will produce one job for each configuration
-        # parameter of type `arduino-platform`, in this case, it
-        # is only 1.
-        arduino-platform: ["arduino:Teensy 4.1"]
-        # This is usually optional but we need to statically define the
-        # FQBN of the boards we want to test for each platform. In the
-        # future the CLI might automatically detect and download the core
-        # needed to compile against a certain FQBN, at that point the
-        # following `include` section will be useless.
-#        include:
-#          # This works like this: when the platformn is "arduino:avr", the
-#          # variable `fqbn` is set to "arduino:avr:uno".
-#          - arduino-platform: "arduino:avr"
-#            fqbn: "arduino:avr:uno"
-
-    # This is the platform GitHub will use to run our workflow,
-    # I picked ubuntu.
     runs-on: ubuntu-latest
 
-    # This is the list of steps this job will run.
+    strategy:
+      matrix:
+        compilation: ['Serial: Blank Sketch']
+        include:
+          - compilation: 'Serial: Blank Sketch'
+            sketch: "$IDE_LOCATION/software/SAT0_Master/SAT0_Master.ino"
+            usb_mode: serial
+    
     steps:
-      # First of all, we clone the repo using the `checkout` action.
       - name: Checkout
-        uses: actions/checkout@main
+        uses: actions/checkout@v3
 
-      # We use the `arduino/setup-arduino-cli` action to install and
-      # configure the Arduino CLI on the system.
-      - name: Setup Arduino CLI
-        uses: arduino/setup-arduino-cli@v1.1.1
-
-
-      # We then install the platform, which one will be determined
-      # dynamically by the build matrix.
-      - name: Install platform
+      - name: Install Arduino IDE
         run: |
-          arduino-cli core update-index
-          arduino-cli core install $
-      # Finally, we compile the sketch, using the FQBN that was set
-      # in the build matrix.
-      - name: Compile Sketch
-        run: arduino-cli compile $ ./software/SAT0_Master --warnings more
+          wget http://downloads.arduino.cc/arduino-$IDE_VERSION-linux64.tar.xz
+          tar xf arduino-$IDE_VERSION-linux64.tar.xz
+          sudo mv arduino-$IDE_VERSION /usr/local/share/arduino
+          sudo ln -s /usr/local/share/arduino/arduino /usr/local/bin/arduino
+          rm arduino-$IDE_VERSION-linux64.tar.xz
+
+      - name: Install Teensyduino
+        run: |
[…]
