+++ .github/workflows/arduino_ci.yml
@@ -13,10 +13,13 @@ jobs:
 
     strategy:
       matrix:
-        compilation: ['Serial: SAT0_Master']
+        compilation: ['Serial: SAT0']
         include:
           - compilation: 'Serial: SAT0_Master'
             sketch: "./software/SAT0_Master/SAT0_Master.ino"
+            usb_mode: serial
+          - compilation: 'Serial: SAT0_Ground'
+            sketch: "./software/SAT0_Master/SAT0_Ground.ino"
             usb_mode: serial
     
     steps:
@@ -38,11 +41,11 @@ jobs:
           sudo ./TeensyduinoInstall.linux64 --dir=/usr/local/share/arduino;
           rm ./TeensyduinoInstall.linux64;
 
-      - name: Remove Teensy Loader Trigger
-        run: sudo python ./.github/workflows/remove_teensyloader.py $IDE_LOCATION/hardware/teensy/avr/platform.txt
+      - name: Install libraries
+        run: arduino-cli lib install "RingBufCPP@1.3"
 
-      - name: Teensy 3.5
-        run: arduino --verify --board teensy:avr:teensy35:usb=${{ matrix.usb_mode }},speed=120,opt=o2std,keys=en-us ${{ matrix.sketch }};
+#      - name: Teensy 3.5
+#        run: arduino --verify --board teensy:avr:teensy35:usb=${{ matrix.usb_mode }},speed=120,opt=o2std,keys=en-us ${{ matrix.sketch }};
 
       - name: Teensy 3.6
         run: arduino --verify --board teensy:avr:teensy36:usb=${{ matrix.usb_mode }},speed=180,opt=o2std,keys=en-us ${{ matrix.sketch }};
