### 2023-04-18 8d7a91d
#21 Add CLI
