+++ .github/workflows/arduino_ci.yml
@@ -41,8 +41,13 @@ jobs:
           sudo ./TeensyduinoInstall.linux64 --dir=/usr/local/share/arduino;
           rm ./TeensyduinoInstall.linux64;
 
+      # We use the `arduino/setup-arduino-cli` action to install and
+      # configure the Arduino CLI on the system.
+      - name: Setup Arduino CLI
+        uses: arduino/setup-arduino-cli@v1.1.1
+
       - name: Install libraries
-        run: arduino-cli lib install "RingBufCPP@1.3"
+        run: arduino-cli lib install --git-url https://github.com/wizard97/Embedded_RingBuf_CPP.git
 
 #      - name: Teensy 3.5
 #        run: arduino --verify --board teensy:avr:teensy35:usb=${{ matrix.usb_mode }},speed=120,opt=o2std,keys=en-us ${{ matrix.sketch }};
