### 2023-04-18 a160e36
#21 Add CLI
