### 2023-04-18 76cd559
#21 Add CLI
