+++ .github/workflows/arduino_ci.yml
@@ -46,8 +46,10 @@ jobs:
       - name: Setup Arduino CLI
         uses: arduino/setup-arduino-cli@v1.1.1
 
+      - name: Set CLI flags
+        run: arduino-cli config set library.enable_unsafe_install true
+
       - name: Install libraries
-        enable_unsafe_install: true
         run: arduino-cli lib install --git-url https://github.com/wizard97/Embedded_RingBuf_CPP.git
 
 #      - name: Teensy 3.5
