+++ .github/workflows/arduino_ci.yml
@@ -47,7 +47,7 @@ jobs:
         uses: arduino/setup-arduino-cli@v1.1.1
 
       - name: Install libraries
-        run: arduino-cli --config-file ./arduino_ci.yml lib install --git-url https://github.com/wizard97/Embedded_RingBuf_CPP.git
+        run: arduino-cli lib install --git-url https://github.com/wizard97/Embedded_RingBuf_CPP.git --enable_unsafe_install true
 
 #      - name: Teensy 3.5
 #        run: arduino --verify --board teensy:avr:teensy35:usb=${{ matrix.usb_mode }},speed=120,opt=o2std,keys=en-us ${{ matrix.sketch }};
