+++ .github/workflows/arduino_ci.yml
@@ -57,6 +57,7 @@ jobs:
           arduino-cli lib install --git-url https://github.com/mikalhart/TinyGPSPlus.git;
           arduino-cli lib install --git-url https://github.com/sparkfun/SparkFun_BQ27441_Arduino_Library.git;
           arduino-cli lib install --git-url https://github.com/jgromes/RadioLib.git;
+          arduino-cli lib install --git-url https://github.com/StuartsProjects/SX12XX-LoRa.git
 
 #      - name: Teensy 3.5
 #        run: arduino --verify --board teensy:avr:teensy35:usb=${{ matrix.usb_mode }},speed=120,opt=o2std,keys=en-us ${{ matrix.sketch }};
