+++ .github/workflows/arduino_ci.yml
@@ -58,7 +58,8 @@ jobs:
           arduino-cli lib install --git-url https://github.com/sparkfun/SparkFun_BQ27441_Arduino_Library.git;
           arduino-cli lib install --git-url https://github.com/jgromes/RadioLib.git;
           arduino-cli lib install --git-url https://github.com/StuartsProjects/SX12XX-LoRa.git;
-          arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_LSM6DS.git
+          arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_LSM6DS.git;
+          arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_BusIO.git
 
 #      - name: Teensy 3.5
 #        run: arduino --verify --board teensy:avr:teensy35:usb=${{ matrix.usb_mode }},speed=120,opt=o2std,keys=en-us ${{ matrix.sketch }};
