+++ .github/workflows/arduino_ci.yml
@@ -61,6 +61,7 @@ jobs:
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_LSM6DS.git;
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_BusIO.git;
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_Sensor.git;
+          arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_LIS3MDL.git;
 
 #      - name: Teensy 3.5
 #        run: arduino --verify --board teensy:avr:teensy35:usb=${{ matrix.usb_mode }},speed=120,opt=o2std,keys=en-us ${{ matrix.sketch }};
