### 2023-04-18 9e2f472
#21 Fix settings.h
