+++ .github/workflows/arduino_ci.yml
@@ -73,4 +73,6 @@ jobs:
         run: arduino --verify --board teensy:avr:teensy36:usb=${{ matrix.usb_mode }},speed=180,opt=o2std,keys=en-us ${{ matrix.sketch }};
 
       - name: Teensy 4.1
-        run: arduino --verify --board teensy:avr:teensy41:usb=${{ matrix.usb_mode }},speed=600,opt=o2std,keys=en-us ${{ matrix.sketch }};
+        run: |
+          arduino-cli lib install --git-url https://github.com/tonton81/WDT_T4.git;
+          arduino --verify --board teensy:avr:teensy41:usb=${{ matrix.usb_mode }},speed=600,opt=o2std,keys=en-us ${{ matrix.sketch }};
