+++ .github/workflows/arduino_ci_ground.yml
@@ -45,13 +45,10 @@ jobs:
         run: |
           arduino-cli lib install --git-url https://github.com/wizard97/Embedded_RingBuf_CPP.git;
           arduino-cli lib install --git-url https://github.com/mikalhart/TinyGPSPlus.git;
-          arduino-cli lib install --git-url https://github.com/sparkfun/SparkFun_BQ27441_Arduino_Library.git;
           arduino-cli lib install --git-url https://github.com/jgromes/RadioLib.git;
           arduino-cli lib install --git-url https://github.com/StuartsProjects/SX12XX-LoRa.git;
-          arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_LSM6DS.git;
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_BusIO.git;
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_Sensor.git;
-          arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_LIS3MDL.git;
 
       - name: ESP32
         run: arduino --verify --board teensy:avr:esp32:usb=${{ matrix.usb_mode }},speed=180,opt=o2std,keys=en-us ${{ matrix.sketch }};
