+++ .github/workflows/arduino_ci_ground.yml
@@ -40,6 +40,7 @@ jobs:
         run: |
           arduino-cli config init;
           arduino-cli config set library.enable_unsafe_install true;
+          arduino-cli core update-index --additional-urls https://espressif.github.io/arduino-esp32/package_esp32_index.json;
 
       - name: Install libraries
         run: |
