+++ .github/workflows/arduino_ci_ground.yml
@@ -52,4 +52,4 @@ jobs:
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_Sensor.git;
 
       - name: ESP32
-        run: arduino --verify --board arduino:avr:heltec_wifi_lora_32_V2,speed=240,opt=o2std,keys=en-us ${{ matrix.sketch }};
+        run: arduino --verify --board espressif32:avr:heltec_wifi_lora_32_V2:usb=${{ matrix.usb_mode }},speed=240,opt=o2std,keys=en-us ${{ matrix.sketch }};
