+++ .github/workflows/arduino_ci_ground.yml
@@ -42,6 +42,7 @@ jobs:
           arduino-cli config set library.enable_unsafe_install true;
           arduino-cli core update-index --additional-urls https://espressif.github.io/arduino-esp32/package_esp32_index.json;
           arduino-cli core install esp32:esp32;
+          
 
       - name: Install libraries
         run: |
