+++ .github/workflows/arduino_ci_ground.yml
@@ -41,7 +41,6 @@ jobs:
           arduino-cli config init;
           arduino-cli config set library.enable_unsafe_install true;
           arduino-cli core update-index --additional-urls https://dl.espressif.com/dl/package_esp32_index.json;
-          arduino-cli core install esp32:esp32;
           
 
       - name: Install libraries
@@ -54,4 +53,4 @@ jobs:
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_Sensor.git;
 
       - name: ESP32
-        run: arduino --verify --board arduino:avr:heltec_wifi_lora_32_V2:usb=${{ matrix.usb_mode }},speed=240,opt=o2std,keys=en-us ${{ matrix.sketch }};
+        run: arduino --verify --board esp32:esp32:esp32:usb=${{ matrix.usb_mode }},speed=240,opt=o2std,keys=en-us ${{ matrix.sketch }};
