+++ .github/workflows/arduino_ci_ground.yml
@@ -40,7 +40,8 @@ jobs:
         run: |
           arduino-cli config init;
           arduino-cli config set library.enable_unsafe_install true;
-          arduino-cli core update-index --additional-urls https://dl.espressif.com/dl/package_esp32_index.json;
+          #arduino-cli core update-index --additional-urls https://dl.espressif.com/dl/package_esp32_index.json;
+          arduino-cli core update-index --additional-urls https://dl.espressif.com/dl/package_esp32_dev_index.json;
           
 
       - name: Install libraries
