+++ .github/workflows/arduino_ci_ground.yml
@@ -42,7 +42,7 @@ jobs:
           arduino-cli config set library.enable_unsafe_install true;
           arduino-cli core update-index --additional-urls https://dl.espressif.com/dl/package_esp32_index.json;
 #          arduino-cli core update-index --additional-urls https://dl.espressif.com/dl/package_esp32_dev_index.json;
-          arduino-cli core install esp32:esp32 -v;
+          arduino-cli core install esp32:esp32;
           
 
       - name: Install libraries
