+++ .github/workflows/arduino_ci_ground.yml
@@ -41,8 +41,7 @@ jobs:
           arduino-cli config init;
           arduino-cli config set library.enable_unsafe_install true;
           arduino-cli core update-index --additional-urls https://dl.espressif.com/dl/package_esp32_index.json;
-#          arduino-cli core update-index --additional-urls https://dl.espressif.com/dl/package_esp32_dev_index.json;
-          arduino-cli core install esp32:esp32;
+          arduino-cli core install esp32:esp32 -v;
           
 
       - name: Install libraries
