+++ .github/workflows/arduino_ci_ground.yml
@@ -52,6 +52,7 @@ jobs:
           arduino-cli lib install --git-url https://github.com/StuartsProjects/SX12XX-LoRa.git;
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_BusIO.git;
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_Sensor.git;
+          arduino-cli lib install --git-url https://github.com/bblanchon/ArduinoJson.git;
 
       - name: ESP32
         run: arduino --verify ${{ matrix.sketch }};
