+++ .github/workflows/arduino_ci_ground.yml
@@ -53,6 +53,7 @@ jobs:
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_BusIO.git;
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_Sensor.git;
           arduino-cli lib install --git-url https://github.com/bblanchon/ArduinoJson.git;
+          arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_SSD1306.git;
 
       - name: ESP32
         run: arduino --verify ${{ matrix.sketch }};
