+++ .github/workflows/arduino_ci_ground.yml
@@ -54,6 +54,7 @@ jobs:
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_Sensor.git;
           arduino-cli lib install --git-url https://github.com/bblanchon/ArduinoJson.git;
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_SSD1306.git;
+          arduino-cli lib install --git-url https://github.com/HelTecAutomation/Heltec_ESP32.git;
 
       - name: ESP32
         run: arduino --verify ${{ matrix.sketch }};
