+++ .github/workflows/arduino_ci_ground.yml
@@ -55,7 +55,7 @@ jobs:
           arduino-cli lib install --git-url https://github.com/bblanchon/ArduinoJson.git;
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_SSD1306.git;
           arduino-cli lib install --git-url https://github.com/HelTecAutomation/Heltec_ESP32.git;
-          arduino-cli lib install --git-url https://github.com/aholinch/sgp4.git;
+          arduino-cli lib install --git-url https://github.com/sparkfun/SparkFun_SGP4_Arduino_Library.git;
 
       - name: ESP32
         run: arduino --verify ${{ matrix.sketch }};
