+++ .github/workflows/arduino_ci_ground.yml
@@ -56,7 +56,8 @@ jobs:
           arduino-cli lib install --git-url https://github.com/adafruit/Adafruit_SSD1306.git;
           arduino-cli lib install --git-url https://github.com/HelTecAutomation/Heltec_ESP32.git;
           arduino-cli lib install --git-url https://github.com/sparkfun/SparkFun_SGP4_Arduino_Library.git;
-          arduino-cli lib install --git-url https://github.com/espressif/arduino-esp32.git;
+          arduino-cli lib install --git-url https://github.com/shurillu/CTBot.git
+          arduino-cli lib install --git-url https://github.com/mobizt/HTTPClientESP32Ex.git;
 
       - name: ESP32
         run: arduino --verify ${{ matrix.sketch }};
