+++ .github/workflows/arduino_ci_ground.yml
@@ -41,7 +41,10 @@ jobs:
           arduino-cli config init;
           arduino-cli config set library.enable_unsafe_install true;
           arduino-cli core update-index --additional-urls https://dl.espressif.com/dl/package_esp32_index.json
-          arduino-cli core install esp32:esp32 --additional-urls - https://raw.githubusercontent.com/espressif/arduino-esp32/gh-pages/package_esp32_index.json
+          arduino-cli core install esp32:esp32 --additional-urls https://espressif.github.io/arduino-esp32/package_esp32_index.json
+          arduino-cli core update-index --additional-urls https://raw.githubusercontent.com/espressif/arduino-esp32/gh-pages/package_esp32_index.json
+          arduino-cli core install esp32:esp32  --additional-urls https://raw.githubusercontent.com/espressif/arduino-esp32/gh-pages/package_esp32_index.json
+          
           
 
       - name: Install libraries
@@ -58,6 +61,7 @@ jobs:
           arduino-cli lib install --git-url https://github.com/sparkfun/SparkFun_SGP4_Arduino_Library.git;
           arduino-cli lib install --git-url https://github.com/shurillu/CTBot.git;
           arduino-cli lib install --git-url https://github.com/jbuszkie/HTTPSRedirect.git;
+          arduino-cli lib install --git-url https://github.com/espressif/arduino-esp32/tree/master/libraries/Wire;
 
       - name: ESP32
         run: arduino --verify ${{ matrix.sketch }};
