+++ .github/workflows/build.yml
@@ -1,0 +1,21 @@
+name: Build examples
+on:
+  push:
+    branches:
+      - master
+      - bugfix/*
+      - feature/*
+  pull_request:
+
+jobs:
+  build-master:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v2
+        with:
+          submodules: 'recursive'
+      - name: esp-idf build
+        uses: espressif/esp-idf-ci-action@latest
+        with:
+          path: './software/SAT0_Ground'
