+++ .github/workflows/build.yml
@@ -1,11 +1,12 @@
 name: Build examples
+
 on:
   push:
     branches:
-      - master
-      - bugfix/*
-      - feature/*
+      - dev
   pull_request:
+    branches:
+      - dev
 
 jobs:
   build-master:
