+++ .github/workflows/arduino_ci_ground.yml
@@ -18,6 +18,11 @@ jobs:
           - compilation: 'Serial: SAT0_Ground'
             sketch: "./software/SAT0_Ground/SAT0_Ground.ino"
             usb_mode: serial
+        board:
+          - fqbn: esp32:esp32:esp32
+            platforms: |
+              - name: esp32:esp32
+                source-url: https://raw.githubusercontent.com/espressif/arduino-esp32/gh-pages/package_esp32_index.json
     
     steps:
       - name: Checkout
@@ -43,9 +48,6 @@ jobs:
           arduino-cli core update-index --additional-urls https://dl.espressif.com/dl/package_esp32_index.json
           arduino-cli core install esp32:esp32 --additional-urls https://espressif.github.io/arduino-esp32/package_esp32_index.json
 
-      - name: Install ESP32 platform dependencies
-        run: pip3 install pyserial
-
       - name: Install libraries
         run: |
           arduino-cli lib install --git-url https://github.com/wizard97/Embedded_RingBuf_CPP.git;
@@ -61,5 +63,13 @@ jobs:
           arduino-cli lib install --git-url https://github.com/shurillu/CTBot.git;
           arduino-cli lib install --git-url https://github.com/jbuszkie/HTTPSRedirect.git;
 
-      - name: ESP32
-        run: arduino --verify --fqbn esp32:esp32:esp32 ${{ matrix.sketch }};
+#      - name: ESP32
+#        run: arduino --verify --fqbn esp32:esp32:esp32 ${{ matrix.sketch }};
+
+      - name: Compile examples
+        uses: arduino/compile-sketches@main
+        with:
+          fqbn: ${{ matrix.board.fqbn }}
+          platforms: ${{ matrix.board.platforms }}
+          enable-deltas-report: true
+          sketches-report-path: ${{ matrix.sketch }};
