+++ .github/workflows/arduino_ci_ground.yml
@@ -69,6 +69,7 @@ jobs:
       - name: Compile examples
         uses: arduino/compile-sketches@main
         with:
-          fqbn: ${{ matrix.board.fqbn }} ${{ matrix.sketch }}
+          fqbn: ${{ matrix.board.fqbn }}
           platforms: ${{ matrix.board.platforms }}
           enable-deltas-report: true
+          sketch-paths: ${{ matrix.sketch }}
