+++ .github/workflows/arduino_ci_ground.yml
@@ -73,4 +73,4 @@ jobs:
           platforms: ${{ matrix.board.platforms }}
           enable-deltas-report: true
           sketch-paths: ${{ matrix.sketch }}
-          cli-compile-flags: -Wno-error
+          cli-compile-flags: --Wno-error
