+++ .github/workflows/arduino_ci_ground.yml
@@ -66,12 +66,13 @@ jobs:
 #      - name: ESP32
 #        run: arduino --verify --fqbn esp32:esp32:esp32 ${{ matrix.sketch }};
 
-      - name: Compile examples
+      - name: Compile ${{ matrix.sketch }}
         uses: arduino/compile-sketches@main
         with:
-          cli-compile-flags: |
-              - --warnings="none"
           fqbn: ${{ matrix.board.fqbn }}
           platforms: ${{ matrix.board.platforms }}
+          cli-compile-flags: |
+            - --build-property
+            - compiler.cpp.extra_flags=-Wno-error=unused-function -Wno-error=unused-variable -Wno-error=deprecated-declarations -Wno-unused-parameter -Wno-sign-compare
           enable-deltas-report: true
           sketch-paths: ${{ matrix.sketch }}
