+++ .github/workflows/arduino_ci_ground.yml
@@ -1,6 +1,12 @@
 name: build SAT0_Ground
 
-on: [push, pull_request]
+on:
+  pull_request:
+    branches:
+      - 'main'
+  push:
+    branches:
+      - 'main'
 
 env:
   IDE_VERSION: 1.8.19
