### 2023-04-18 #5147 @ronyronen
@federeghe @prashjha Hi, I've added a github action for CI for both, the SAT0_Master and SAT0_Ground. Both are compiled using the CI. Do let me now if further information is required. A new release 4.0.10 has been published.
