+++ .github/workflows/arduino_ci_master.yml
@@ -1,9 +1,6 @@
 name: build SAT0_Master
 
 on:
-  pull_request:
-    branches:
-      - 'main'
   push:
     branches:
       - 'main'
