+++ .github/workflows/arduino_ci_ground.yml
@@ -1,9 +1,6 @@
 name: build SAT0_Ground
 
 on:
-  pull_request:
-    branches:
-      - 'main'
   push:
     branches:
       - 'main'
