### 2023-04-24 #5147 @nachootal
Hi @ronyronen ,

my feedback is the following:

- Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.

Data is available in the software folder. From the description in the paper I can't really assess that there's a library because the naming is a bit misleading. I see the 3 parts of software that the paper refers but I don't see clearly where the library is available

- Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

I think this part was also asked by @federeghe and you mentioned you'd be providing a GitHub action. Thanks for that.

- Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I find the example needs to come probably by linking to a paper or description of a STATLLA-2B satellite so that it can be checked that the architecture matches and that it can be installed there and how.

- Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

If the object of the paper is the library, there's no documentation for the API of that library that I could find

- Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Same as above.
