### 2023-04-25 #5147 @ronyronen
Hi, @nachootal, Thanks for your review.

I'll work on the issues and will update the thread.

> Data is available in the software folder. From the description in the paper I can't really assess that there's a library because the naming is a bit misleading. I see the 3 parts of software that the paper refers but I don't see clearly where the library is available

Please note that this was raised in issue [#14](https://github.com/kcglab/satllazero/issues/14) and has been fixed according to 2nd reviewer.

> Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

Actions is implemented. Please check action results in version 4.0.11 of the repo.

> I find the example needs to come probably by linking to a paper or description of a STATLLA-2B satellite so that it can be checked that the architecture matches and that it can be installed there and how.

We refer the reader to SATLLA-2B tracking website and included all the materials to build it includes the hardware sketch and BOM.

> If the object of the paper is the library, there's no documentation for the API of that library that I could find

The aim of the paper/library is not an API but to build a fully functional pico-satellite model based on a proven design that is fully functional in space.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

We did include the github actions to build the software. Going forward we do want to enhance this repo together with the community and include more features. I'll check further on manual steps test and update the thread.

Thanks again for your review.
