### 2023-04-26 #5147 @nachootal
Hi @ronyronen ,

Thanks for such a fast response :)

I ticked already some of the comments I pointed because they were, as you said, covered in other issues.

I clarify some of my comments yet that I leave as suggestions only for you to estimate.

- I see the link to SATLLA-0 on line 67. My suggestion was adding the link in every mention of the SATLLA-0. I ticked the requirement in because it's certainly linked. I just think for a newbie like me it would probably make sense to find this link early enough in the document.
- The API description in the projects is missing. I understand this point of the revision as if there's an HTML version of the API like the one you'd find - in this concrete case - in pypi if you'd publish that piece of software as a library. From your description in the summary, I understand that there's two parts: Hardware and Software. The Hardware part is so that the thing can be built, the software part is, somehow, to make it run. When you mention that this software is divided in three libraries I understand that these components can be independently interfaced and used. Perhaps this is bias on my side that made me expect this and this is why I found interesting that the API for each of those libraries should be documented in case someone wants to use them as part of their projects. I don't know if it makes sense saying that these libraries are meant to be used as part of the project only because they're custom to this certain scope. (As I said, this is probably my bias so evaluate if it makes sense stating this clearer).

Thanks again for your attention and interesting work,
Nacho
