### 2023-04-28 #5147 @ronyronen
Hi @nachootal Thanks for your comments.

I'v update the paper doc to include the link to the repository early in the paper as you suggested.

To be clear about the software, although combining all of the components would result in a functioning satellite, each individual component can also function independently. For example, the flight software (SAT0_Master) can be used as a standalone, and the experiments from SAT0_OBC can be used independently as well. Do you think I should emphasize this in the paper? Thanks again.
