+++ paper/paper.md
@@ -27,9 +27,9 @@ bibliography: paper.bib
 # Summary
 Nanosatellites are becoming a preferred platform for testing innovative technologies and conducting academic research in space. Flight Software (FSW) is software that runs on a processor embedded in the satellite body [@miranda2019comparative]. The software is responsible for managing satellite activity, data processing, and satellite health and safety, thus enabling the satellite to perform all actions necessary to achieve its scientific objective.
 
-SATLLA-0 is an open source platform for building and experimenting with laboratory nanosatellites for academics and school students. The platform is based on the flight software of the SATLLA-2B satellite, developed and built by the Nanosatellite Research Laboratory of Ariel University. The SATLLA-2B pico-satellite was successfully launched on January 13, 2022 on board a Falcon 9 launch vehicle. Since launched, SATLLA-2B is fully operational and its messages are being received on a daily basis by a large set of TinyGS.com users.
+[SATLLA-0](https://github.com/kcglab/satllazero) is an open source platform for building and experimenting with laboratory nanosatellites for academics and school students. The platform is based on the flight software of the [SATLLA-2B](https://tinygs.com/satellite/SATLLA-2B) satellite, developed and built by the Nanosatellite Research Laboratory of Ariel University. The SATLLA-2B pico-satellite was successfully launched on January 13, 2022 on board a Falcon 9 launch vehicle. Since launched, SATLLA-2B is fully operational and its messages are being received on a daily basis by a large set of TinyGS.com users.
 
-SATLLA-0 is divided into three main libraries, with the main library, SAT0_Master, responsible for the ongoing operation of the satellite, the SAT0_OBC library responsible for the operation of the satellite's research objective, and the SAT0_Ground library responsible for transmitting telecommand to the satellite and receiving satellite transmissions and routing the data for further analysis.
+SATLLA-0 is divided into three main libraries, with the main library, [SAT0_Master](https://github.com/kcglab/satllazero/tree/main/software/SAT0_Master), responsible for the ongoing operation of the satellite, the [SAT0_OBC](https://github.com/kcglab/satllazero/tree/main/software/SAT0_OBC) library responsible for the operation of the satellite's research objective, and the [SAT0_Ground](https://github.com/kcglab/satllazero/tree/main/software/SAT0_Ground) library responsible for transmitting telecommand to the satellite and receiving satellite transmissions and routing the data for further analysis.
 
 
 # Statement of need
