### 2023-05-08 967174e
#29 Emphasize that the lib can be used independently
