+++ software/SAT0_Ground/settings.h
@@ -206,9 +206,9 @@ bool REPLY_ACK = false;
 #define LORA_433_RST_PIN 23
 #define LORA_433_DIO0_PIN 26
 #elif (DEVCIE == BOARD_T_3V16) // TTGO T3_V1.6
-#define LORA_433_NSS_PIN 4
-#define LORA_433_RST_PIN 3
-#define LORA_433_DIO0_PIN 2
+#define LORA_433_NSS_PIN 18
+#define LORA_433_RST_PIN 14
+#define LORA_433_DIO0_PIN 26
 #endif
 
 #define LORA_433_CR_4_5 5
