### 2023-06-07 #5147 @ronyronen
@federeghe @prashjha @nachootal We are delighted to introduce the SATLLA-0 Evaluation Module (EVM) board. This board enhances the testing and evaluation capabilities of SATLLA-0. Equipped with a diverse range of satellite sensors, the EVM board facilitates comprehensive functionality experiments. It enables thorough analysis and optimization of various aspects such as thermal management, power optimization, and communication protocols. [evm_board](https://github.com/kcglab/satllazero/blob/main/hardware/evm_board/SATLLA_0_EVM_Board.jpeg).

Nevertheless, we wanted to hear from you about the review process. Can you please reply to my previous request?
