+++ paper/paper.md
@@ -12,6 +12,7 @@ authors:
     orcid: 0000-0002-1300-5236
     affiliation: "1"
   - name: Michael Britvin
+    orcid: 0009-0003-5492-4014
     affiliation: "2"
   - name: Boaz Ben Moshe
     orcid: 0000-0002-1580-5421
