### 2023-06-20 #5147 @ronyronen
> https://github.com/kcglab/satllazero

Could you confirm if you are referring to this link: https://zenodo.org/record/8060399?
