### 2023-06-20 #5147 @prashjha
Yes, but the title of zenodo needs to match the title of your JOSS paper; see examples in these closed issues: https://github.com/openjournals/joss-reviews/issues?q=label%3Areview+is%3Aclosed
