### 2023-06-21 #5147 @prashjha
Hi, @ronyronen, see if these suggestions make sense for your draft:

- Use bullet points or numbered list for items in Lines 29 - 33
- Figure 1 is missing caption and the label that says 'Figure 1'. See papers published in JOSS for examples. Same is true for figure 2.
- Inconsistent references to figures. E.g., you refer to figure 1 as Fig. 1 and figure 2 as Figure 2. You can see some example papers in JOSS to also use hyperlink.
- I will recommend creating a list for items in Lines 51 and 57 and highlighting `Initialization` and `Operation` by considering bold font weight.

Let me know when you are finished editing the draft.
