### 2023-06-22 kcglab/satllazero#35 @ronyronen
Paper review issues
https://github.com/openjournals/joss-reviews/issues/5147#issuecomment-1601281155
