+++ paper/paper.md
@@ -26,42 +26,39 @@ date: 19 December 2022
 bibliography: paper.bib
 ---
 # Summary
-Nanosatellites are becoming a preferred platform for testing innovative technologies and conducting academic research in space. Flight Software (FSW) is software that runs on a processor embedded in the satellite body [@miranda2019comparative]. The software is responsible for managing satellite activity, data processing, and satellite health and safety, thus enabling the satellite to perform all actions necessary to achieve its scientific objective.
+Nanosatellites becoming a preferred platform for testing innovative technologies and conducting academic research in space. Flight Software (FSW) is a software program that operates on a processor embedded within the satellite structure [@miranda2019comparative]. It assumes the crucial responsibility of managing various aspects of satellite operations, including activity control, data processing, and ensuring satellite health and safety. By fulfilling these tasks, the FSW empowers the satellite to execute all actions essential for achieving its scientific objectives.
 
-[SATLLA-0](https://github.com/kcglab/satllazero) is an open source platform for building and experimenting with laboratory nanosatellites for academics and school students. The platform is based on the flight software of the [SATLLA-2B](https://tinygs.com/satellite/SATLLA-2B) satellite, developed and built by the Nanosatellite Research Laboratory of Ariel University. The SATLLA-2B pico-satellite was successfully launched on January 13, 2022 on board a Falcon 9 launch vehicle. Since launched, SATLLA-2B is fully operational and its messages are being received on a daily basis by a large set of TinyGS.com users.
+[SATLLA-0](https://github.com/kcglab/satllazero) is an open-source platform designed for academic institutions and school students to build and explore laboratory nanosatellites. It is based on the flight software utilized by the [SATLLA-2B](https://tinygs.com/satellite/SATLLA-2B) nanosatellite, which was developed and constructed by the Nanosatellite Research Laboratory of Ariel University. The successful launch of the SATLLA-2B nanosatellite took place on January 13, 2022, aboard a Falcon 9 launch vehicle. Since its launch, SATLLA-2B has been operating seamlessly, with its transmissions being received daily by a wide range of users on tinyGS.com.
 
-SATLLA-0 is divided into three main libraries, with the main library, [SAT0_Master](https://github.com/kcglab/satllazero/tree/main/software/SAT0_Master), responsible for the ongoing operation of the satellite, the [SAT0_OBC](https://github.com/kcglab/satllazero/tree/main/software/SAT0_OBC) library responsible for the operation of the satellite's research objective, and the [SAT0_Ground](https://github.com/kcglab/satllazero/tree/main/software/SAT0_Ground) library responsible for transmitting telecommand to the satellite and receiving satellite transmissions and routing the data for further analysis.
-
+SATLLA-0 is divided into three main libraries, with the main library, [SAT0_Master](../blob/master/software/SAT0_Master), responsible for the ongoing operation of the satellite, the [SAT0_OBC](../blob/master/software/SAT0_OBC) library responsible for the operation of the satellite's research objective, and the [SAT0_Ground](../blob/master/software/SAT0_Ground) library responsible for transmitting telecommand to the satellite and receiving satellite transmissions and routing the data for further analysis.
 
 # Statement of need
 Building a nanosatellite and developing flight software is challenging for newcomers to the space (i.e., new-space) field due to the lack of literature and high cost. There are open-source frameworks for space projects developed by space agencies, universities, or commercial companies to serve as reference designs and encourage code reuse in their future projects [@jamie2017basic]. These frameworks have not yet gained significant adoption outside of their host organizations [@miranda2019comparative] for the following reasons:
-- Flight software is tailored to a satellite of a specific design and is not transferable to another design.
-- Flight software is not based on open source software or the entire flight software has not been released as open source software.
[…]
