+++ paper/paper.md
@@ -30,7 +30,7 @@ Nanosatellites becoming a preferred platform for testing innovative technologies
 
 [SATLLA-0](https://github.com/kcglab/satllazero) is an open-source platform designed for academic institutions and school students to build and explore laboratory nanosatellites. It is based on the flight software utilized by the [SATLLA-2B](https://tinygs.com/satellite/SATLLA-2B) nanosatellite, which was developed and constructed by the Nanosatellite Research Laboratory of Ariel University. The successful launch of the SATLLA-2B nanosatellite took place on January 13, 2022, aboard a Falcon 9 launch vehicle. Since its launch, SATLLA-2B has been operating seamlessly, with its transmissions being received daily by a wide range of users on tinyGS.com.
 
-SATLLA-0 is divided into three main libraries, with the main library, [SAT0_Master](../blob/master/software/SAT0_Master), responsible for the ongoing operation of the satellite, the [SAT0_OBC](../blob/master/software/SAT0_OBC) library responsible for the operation of the satellite's research objective, and the [SAT0_Ground](../blob/master/software/SAT0_Ground) library responsible for transmitting telecommand to the satellite and receiving satellite transmissions and routing the data for further analysis.
+SATLLA-0 is divided into three main libraries, with the main library, [SAT0_Master](../software/SAT0_Master), responsible for the ongoing operation of the satellite, the [SAT0_OBC](../software/SAT0_OBC) library responsible for the operation of the satellite's research objective, and the [SAT0_Ground](../software/SAT0_Ground) library responsible for transmitting telecommand to the satellite and receiving satellite transmissions and routing the data for further analysis.
 
 # Statement of need
 Building a nanosatellite and developing flight software is challenging for newcomers to the space (i.e., new-space) field due to the lack of literature and high cost. There are open-source frameworks for space projects developed by space agencies, universities, or commercial companies to serve as reference designs and encourage code reuse in their future projects [@jamie2017basic]. These frameworks have not yet gained significant adoption outside of their host organizations [@miranda2019comparative] for the following reasons:
