### 2023-06-22 #5147 @kyleniemeyer
Hello @ronyr, a few final things to fix before publishing:

- the Zenodo archive has an extraneous author/contributor who isn't on the paper author list. I assume that is a mistake?
- The first sentence of the paper is missing a word: "Nanosatellites **are** becoming..."
- The Miranda et al. reference appears to be missing a space in the title (between "space" and "nanosatellites"
