+++ paper/paper.bib
@@ -18,7 +18,7 @@
 }
 
 @article{miranda2019comparative,
-  title={A comparative survey on flight software frameworks for ‘new space’nanosatellite missions},
+  title={A Comparative Survey on Flight Software Frameworks for ‘New Space’ Nanosatellite Missions},
   author={Miranda, Danilo Jos{\'e} Franzim and Ferreira, Maur{\'\i}cio and Kucinskis, Fabricio and McComas, David},
   journal={Journal of Aerospace Technology and Management},
   volume={11},
