+++ README.md
@@ -81,7 +81,7 @@ This program is released as open source software under the terms of [GPL3 Licens
 2. SATLLA 2 DIY KIT Assembly [iFixIt Build Instructions](https://www.ifixit.com/Guide/SATLLA+2+DIY+KIT+Assembly/147004)
 3. A good starting point is to use a Weather Balloons: [Extreme Long-RAnge Wi-Fi](https://www.youtube.com/watch?v=0xc7XjHUJkM&t=41s)
 
-#Citation
+# Citation
 If you'd like to cite us in a project or publication, please include a reference to the JOSS paper:
 ```
 @article{Ronen_SATLLA0_A_Flight_2023,
