### 2023-06-25 98bf32b
Update PINs for the Teensy3.2 (EVM)
