### 2023-06-26 #5147 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
